services/tts/chatterbox-turbo/requirements.txt
fastapi
uvicorn
python-dotenv
torchaudio
chatterbox-tts
scipy